# docs/openapi_skeleton.yaml
openapi: 3.1.0
info:
  title: Gestionale Collaboratori e Progetti
  description: "\n    Sistema gestionale completo per:\n    - \U0001F465 Collaboratori\
    \ (anagrafica, ruoli, tariffe)\n    - \U0001F4CB Progetti (budget, ente attuatore,\
    \ stato)\n    - \U0001F4DD Incarichi (assegnazioni ai progetti con mansioni)\n\
    \    - ⏰ Presenze (rilevazioni orarie con validazioni)\n    - \U0001F4C4 Contratti\
    \ (collaborazione, P.IVA, subordinato)\n    - \U0001F4B0 Rendicontazione (FSE+,\
    \ Fondimpresa, PNRR)\n    "
  version: 1.0.0
paths:
...
  /api/v1/contracts/templates:
    get:
      tags:
      - Contracts
      summary: Get Templates
      description: Lista tutti i template contratti
      operationId: get_templates_api_v1_contracts_templates_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ContractTemplate'
                type: array
                title: Response Get Templates Api V1 Contracts Templates Get
    post:
      tags:
      - Contracts
      summary: Create Template
      description: Crea un nuovo template contratto
      operationId: create_template_api_v1_contracts_templates_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractTemplateCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractTemplate'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/contracts/templates/{template_id}:
    delete:
      tags:
      - Contracts
      summary: Delete Template
      description: Elimina un template
      operationId: delete_template_api_v1_contracts_templates__template_id__delete
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ContractTemplate:
      properties:
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        template_type:
          type: string
          title: Template Type
        content:
          type: string
          title: Content
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - name
      - template_type
      - content
      - id
      - created_at
      title: ContractTemplate
    ContractTemplateCreate:
      properties:
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        template_type:
          type: string
          title: Template Type
        content:
          type: string
          title: Content
      type: object
      required:
      - name
      - template_type
      - content
      title: ContractTemplateCreate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError